+++ internal/provider/resource_sdwan_rewrite_rule_policy_definition.go
@@ -59,33 +59,26 @@ func (r *RewriteRulePolicyDefinitionResource) Metadata(ctx context.Context, req 
 func (r *RewriteRulePolicyDefinitionResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: helpers.NewAttributeDescription("This resource can manage a Rewrite Rule policy definition.").String,
+		MarkdownDescription: helpers.NewAttributeDescription("This resource can manage a Rewrite Rule Policy Definition .").String,
 
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
-				MarkdownDescription: "The id of the policy definition",
+				MarkdownDescription: "The id of the object",
 				Computed:            true,
 				PlanModifiers: []planmodifier.String{
 					stringplanmodifier.UseStateForUnknown(),
 				},
 			},
 			"version": schema.Int64Attribute{
-				MarkdownDescription: "The version of the policy definition",
+				MarkdownDescription: "The version of the object",
 				Computed:            true,
 			},
 			"name": schema.StringAttribute{
-				MarkdownDescription: "The name of the policy definition",
+				MarkdownDescription: helpers.NewAttributeDescription("The name of the policy definition").String,
 				Required:            true,
 			},
-			"type": schema.StringAttribute{
-				MarkdownDescription: "The policy defintion type",
-				Computed:            true,
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.UseStateForUnknown(),
-				},
-			},
 			"description": schema.StringAttribute{
-				MarkdownDescription: "The description of the policy definition",
+				MarkdownDescription: helpers.NewAttributeDescription("The description of the policy definition").String,
 				Required:            true,
 			},
 			"rules": schema.ListNestedAttribute{
@@ -115,7 +108,7 @@ func (r *RewriteRulePolicyDefinitionResource) Schema(ctx context.Context, req re
 								int64validator.Between(0, 63),
 							},
 						},
-						"layer2cos": schema.Int64Attribute{
+						"layer2_cos": schema.Int64Attribute{
 							MarkdownDescription: helpers.NewAttributeDescription("Layer2 CoS").AddIntegerRangeDescription(0, 7).String,
 							Optional:            true,
 							Validators: []validator.Int64{
@@ -139,7 +132,7 @@ func (r *RewriteRulePolicyDefinitionResource) Configure(_ context.Context, req r
 }
 
 func (r *RewriteRulePolicyDefinitionResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
-	var plan RewriteRule
+	var plan RewriteRulePolicyDefinition
 
 	// Read plan
 	diags := req.Plan.Get(ctx, &plan)
@@ -153,7 +146,7 @@ func (r *RewriteRulePolicyDefinitionResource) Create(ctx context.Context, req re
 	// Create object
 	body := plan.toBody(ctx)
 
-	res, err := r.client.Post("/template/policy/definition/rewriterule", body)
+	res, err := r.client.Post("/template/policy/definition/rewriterule/", body)
 	if err != nil {
 		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to configure object (POST), got error: %s, %s", err, res.String()))
 		return
@@ -161,7 +154,6 @@ func (r *RewriteRulePolicyDefinitionResource) Create(ctx context.Context, req re
 
 	plan.Id = types.StringValue(res.Get("definitionId").String())
 	plan.Version = types.Int64Value(0)
-	plan.Type = types.StringValue(plan.getType())
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Create finished successfully", plan.Name.ValueString()))
 
@@ -170,15 +162,10 @@ func (r *RewriteRulePolicyDefinitionResource) Create(ctx context.Context, req re
 }
 
 func (r *RewriteRulePolicyDefinitionResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
-	var state, oldState RewriteRule
+	var state RewriteRulePolicyDefinition
 
 	// Read state
 	diags := req.State.Get(ctx, &state)
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-	diags = req.State.Get(ctx, &oldState)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
@@ -196,7 +183,6 @@ func (r *RewriteRulePolicyDefinitionResource) Read(ctx context.Context, req reso
 	}
 
 	state.fromBody(ctx, res)
-	state.updateVersions(ctx, oldState)
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Read finished successfully", state.Name.ValueString()))
 
@@ -205,7 +191,7 @@ func (r *RewriteRulePolicyDefinitionResource) Read(ctx context.Context, req reso
 }
 
 func (r *RewriteRulePolicyDefinitionResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	var plan, state RewriteRule
+	var plan, state RewriteRulePolicyDefinition
 
 	// Read plan
 	diags := req.Plan.Get(ctx, &plan)
@@ -238,7 +224,6 @@ func (r *RewriteRulePolicyDefinitionResource) Update(ctx context.Context, req re
 	} else {
 		tflog.Debug(ctx, fmt.Sprintf("%s: No changes detected", plan.Name.ValueString()))
 	}
-
 	plan.Version = types.Int64Value(state.Version.ValueInt64() + 1)
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Update finished successfully", plan.Name.ValueString()))
@@ -248,7 +233,7 @@ func (r *RewriteRulePolicyDefinitionResource) Update(ctx context.Context, req re
 }
 
 func (r *RewriteRulePolicyDefinitionResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
-	var state RewriteRule
+	var state RewriteRulePolicyDefinition
 
 	// Read state
 	diags := req.State.Get(ctx, &state)