+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `sdwan_traffic_data_policy_definition` resource and data source
 - Add `set_parameters` to `sdwan_acl_policy_definition` resource and data source
 - BREAKING CHANGE: Remove `dscp` and `next_hop` attributes of `sdwan_acl_policy_definition` resource and data source
+- BREAKING CHANGE: Rename `layer2cos` attribute of `sdwan_rewrite_rule_policy_definition` resource and data source to `layer2_cos`
 
 ## 0.2.1